Hashage du code dans la mémoire pour être RGPD compliant
